Initialize sponsor wallets in e2e test
