+++ src/state/state.ts
@@ -2,16 +2,7 @@ import type { BigNumber } from 'ethers';
 import { produce, type Draft } from 'immer';
 
 import type { Config } from '../config/schema';
-import type {
-  DapiName,
-  PrivateKey,
-  DecodedDataFeed,
-  ChainId,
-  SignedData,
-  DataFeedId,
-  ProviderName,
-  DapiName,
-} from '../types';
+import type { DapiName, PrivateKey, DecodedDataFeed, ChainId, SignedData, DataFeedId, ProviderName } from '../types';
 
 interface GasState {
   gasPrices: { price: BigNumber; timestampMs: number }[];
